+++ fiddles/nativeScript/fiddle-0014-SiteWrapper/sitewrapper/app/main-page.js
@@ -1,0 +1,29 @@
+/*
+In NativeScript, a file with the same name as an XML file is known as
+a code-behind file. The code-behind is a great place to place your view
+logic, and to set up your page’s data binding.
+*/
+
+/*
+NativeScript adheres to the CommonJS specification for dealing with
+JavaScript modules. The CommonJS require() function is how you import
+JavaScript modules defined in other files.
+*/ 
+var createViewModel = require("./main-view-model").createViewModel;
+var utilsModule = require("tns-core-modules/utils/utils");
+
+
+function pageLoaded(args) {
+  var page = args.object;
+  page.bindingContext = { };
+  utilsModule.openUrl("http://fiddle.sh/");
+}
+
+/*
+Exporting a function in a NativeScript code-behind file makes it accessible
+to the file’s corresponding XML file. In this case, exporting the onNavigatingTo
+function here makes the navigatingTo="onNavigatingTo" binding in this page’s XML
+file work.
+*/
+exports.pageLoaded = pageLoaded;
+